bus/pci: fix bug with mmapped PCI resources
pci_scan_one() corrupts the dev->mem_resource[] field for the
already-mapped-in-userspace registered PCI devices. A workaround is to
override the mem_resource[] field either when the device is not bound to
a userspace driver or the userspace driver does not require PCI resource
mapping.
Bug description: https://bugs.dpdk.org/show_bug.cgi?id=85

Signed-off-by: Nikos Dragazis <ndragazis@outlook.com.gr>