vfio: map all PCI BARs in 2MB-aligned virtual addresses
This is a workaround to a restriction of SPDK env_dpdk library. The
vtophys map requires the registered addresses to be 2MB aligned. This
patch enables registering a mapped BAR to the vtophys map of SPDK in
case of VFIO no-IOMMU mode. This is useful in case of peer-to-peer DMA
from a PCI device to a PCI BAR of another device. This is the case of
the SPDK vhost target with vvu transport.

Signed-off-by: Nikos Dragazis <ndragazis@outlook.com.gr>